Add support for a size-specified allocation API to memory-manager
Adds a new API where you tell the manager the sizes of the
objects during free, so it doesn't need to track it internally.  This
is for use with in situ string allocation and moving ObjectData
allocation into the shared free lists.  There are a few things still
needing cleanup in the API here still (e.g. smartMallocSizeBig has to
add sizeof(SmallNode) just so it can be subtracted), but I plan to do
it on top of this diff stack to isolate it from causing any perf
impact.

Differential Revision: D953828