Delete a bunch of dead code in TranslatorX64
I just went through the header and grepped for each member function
name, and it turns out a bunch of them were declared and not
implemented, or were implemented but not used.

This clarifies the picture of what assembler-coupled functionality
remains on TranslatorX64:

1. Func prologues and their transitive dependencies (including
unfortunate stuff like a duplicated implementation of emitIncRef).

2. One-time-use stub emitters.

3. Syncing VM registers.

This makes the project seems a little tamer. Func prologues are
definitely the biggest part of it.

Differential Revision: D921995