remove invalid assert in http_protocol
http_protocol was trying to assert that every body chunk was NULL terminated.  libevent_server *thought* it was doing this by enlarging the buffer and adding a 0.  POST body data can be binary though (images, video, etc) so if any code was actually relying on the 0 for termination, it was likely wrong.

The proxygen server will not add 0 termination to body chunks, so this assert has to go.

Reviewed By: @markw65

Differential Revision: D923553