add ftp library
This is a good one. It doesn't need an external library and found a ton of corner cases I screwed up. It also got my stream handling much more solid.

I had to wrap other streams as they came into the extension and then unwrap them on the way out

Reviewed By: @sgolemon

Differential Revision: D945250