import yaml extension
I'm putting this in the facebook directory since I don't want to add yaml as a depedancy for the OSS version. This isn't even in the public version of php.net it is in pear, but I want this to set the precident of how to import pear extensions (until @sgolemon makes us a better way). At least this way we get the unit tests.

Reviewed By: @sgolemon

Differential Revision: D942072