Expand JIT::relaxGuards to work during normal translation
This is the last major piece of guard relaxation. It adds support for
reflowing the relaxed guard types, meaning it can be applied to a trace without
leaving it in an inconsistent state. The main benefit of this is that it allows
us to loosen/eliminate unnecessary predictions in the region translator.

Reviewed By: @ottoni

Differential Revision: D943543